De-duplicate the tests
